2005-09-09 08:30 UTC+0200 Lorenzo Fiorini <lorenzo_fiorini/at/teamwork.it>   * gtk/Makefile   * gtk/Makefile.contrib   * gtk/build.sh     * Makefile renamed to Makefile.linux and Makefile.contrib renamed Makefile   * gtk/commond.c   * gtk/control.c   * gtk/menu_c.c   * gtk/window.c   * source/freeimg.c   * source/message.c   * source/nice.c   * source/window.c   * source/xml/xmlparse.c   * source/procmisc/arr2str.c     * fixed warnings for mingw32   * gtk/wprint.c     * added #ifndef __MINGW32__   * gtk/hbrowse.prg   * gtk/hdialog.prg   * gtk/menu.prg     * fixed warnings for -W
git-svn-id: https://svn.code.sf.net/p/hwgui/code/trunk/hwgui@446 66345640-f411-406e-acba-706ed0d45c9c

+++ gtk/control.c
@@ -1,5 +1,5 @@
 /*
- * $Id: control.c,v 1.10 2005-09-07 05:06:45 alkresin Exp $
+ * $Id: control.c,v 1.11 2005-09-09 06:30:20 lf_sfnet Exp $
  *
  * HWGUI - Harbour Linux (GTK) GUI library source code:
  * Widget creation functions
@@ -248,7 +248,7 @@ HB_FUNC( HWG_COMBOSETARRAY )
    char * cItem;
    int i;
 
-   for( i=0; i<pArr->item.asArray.value->ulLen; i++ )
+   for( i=0; (ULONG)i<pArr->item.asArray.value->ulLen; i++ )
    {
       cItem = g_locale_to_utf8( hb_itemGetCPtr( pArr->item.asArray.value->pItems + i ),-1,NULL,NULL,NULL );
       glist = g_list_append( glist, cItem );
